+++ rest_framework_json_api/filters/__init__.py
@@ -1,0 +1,2 @@
+from .sort import JSONAPIOrderingFilter  # noqa: F401
+from .filter import JSONAPIDjangoFilter  # noqa: F401
